fix bugs and add new const to html report

+++ templates/assembly_stats_global.py
@@ -115,13 +115,13 @@ def main(sample_id, assembler, assembly, read_mapping_stats, min_len, n_target):
                     "contigs": len(contigs),
                     "basepairs": sum(contigs),
                     "max_contig_size": max(contigs) if len(contigs) > 0 else 0,
-                    "N{}".format(n_target*100): n50_contigs,
+                    "N{}".format(int(n_target*100)): n50_contigs,
                     "mapped_reads": mapped_reads},
                 "filtered": {
                         "min_len": min_len,
                         "contigs": len(contigs_over_min_len),
                         "basepairs": sum(contigs_over_min_len),
-                        "N{}".format(n_target*100): n50_contigs_over_min_len}
+                        "N{}".format(int(n_target*100)): n50_contigs_over_min_len}
                 }
 
         json_report.write(json.dumps(json_dic, separators=(",", ":")))
